Window form added, changed how screens are drawn
Created a 'window' form which has the other forms added to it as controls. this means we are not hiding and showing new forms but appending forms to a 'base' form.
Issue with resizing screens as current to redraw the screen on resize, we have to clear and redraw. This means clearing out text boxes which is not ideal.

Co-Authored-By: O. <22729676+OliverOsborne@users.noreply.github.com>