Redrawing projForm panels on resize
Instead of making a new projForm on resize, now DrawPanels method is called to draw all the projects to the screen. Still slow however this solves the issue of data being lost on resize.
This will have to be looked at on a form-by-form basis to decide what needs redrawing and how to store data (if required)