BasePanel drawing function is drawPanels
Changed drawing panel logic so that DrawPanel.BasePanels is a method that takes a panel to draw to and a list of datasets. This creates panels based on how many datasets there are and adds them to the provided panel. Then, ProjectPanels is called which takes the panel to add controls to (the panel just made in basePanels)  and a dataset, it then adds the items in the dataset to the panel as clickable project panels.
very buggy