notification skeleton
added notifications and notifType into db, new SqlNotifcation class to handle these.
notification table stores users and the projects to notify them about. notifType stores what info to send to users in notification table.
needed to allow projects to be made private and users to be allowed to request access